fonction de tir de la team Underscore
