+++ py/team_underscore/shoot/readme
@@ -1,0 +1,14 @@
+******************************
+********README****************
+******************************
+
+Dossier de la fonction shoot
+
+Nom des fonctions:
+
+- 	leftShoot() #fini
+- 	rightShoot() #fini
+- 	sideLeftShoot() #fini
+- 	sideRightShoot() #fini
+
+Appelez Quentin en cas d'erreur :)
