--- shoot/readme
@@ -1,7 +1,0 @@
-******************************
-********README****************
-******************************
-
-Dossier de la fonction shoot
-
-
